requirements.txt
# requirements.txt
Flask
Flask-Cors
Flask-Bcrypt
psycopg2-binary
PyJWT
